+++ src/datatype/create_struct.rs
@@ -48,7 +48,7 @@ pub unsafe fn generate_c_struct(
         offset += size + padding;
         size
       }
-      RsArgsValue::I64(number) => {
+      RsArgsValue::I64(number) | RsArgsValue::BigInt(number) => {
         let (size, align) = get_size_align::<c_longlong>();
         let padding = (align - (offset % align)) % align;
         field_ptr = field_ptr.offset(padding as isize);